+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 1. [13510](https://github.com/influxdata/influxdb/pull/13510): Add org name to dahboard page title 
 1. [13520](https://github.com/influxdata/influxdb/pull/13520): Add cautioning to bucket renaming
 1. [13560](https://github.com/influxdata/influxdb/pull/13560): Add option to generate all access token in tokens tab
+1. [13601](https://github.com/influxdata/influxdb/pull/13601): Add option to generate read/write token in tokens tab
 
 ### Bug Fixes